minor fixes and cosmetic changes
- fixed SN76496 crashing in emutest/audemutest
- verified Virtual Boy VSU sample rate
- fixed YM2151 per-channel clipping
- changed some variables from signed to unsigned
- eliminated some GCC warnings

+++ emu/cores/okim6258.c
@@ -161,7 +161,7 @@ static void compute_tables(void)
 	for (step = 0; step <= 48; step++)
 	{
 		/* compute the step value */
-		int stepval = floor(16.0 * pow(11.0 / 10.0, (double)step));
+		int stepval = (int)floor(16.0 * pow(11.0 / 10.0, step));
 
 		/* loop over all nibbles and compute the difference */
 		for (nib = 0; nib < 16; nib++)
